Add convenient C++ wrapper classes for the CUDD library:
  * Support for automatic referencing/dereferening.
  * ADD/BDD type safety.
  * Operator overloading.
  * C++11 move semantics.

There are C++ wrapper classes distributed with the CUDD library, but they
contain a lot of functionality not used by Ymer, do not fit the style of the
Ymer code, and make use of virtual destructors.  Polymorphism is not terribly
useful for ADDs/BDDs, at least not in the Ymer source code, so we prefer to
avoid the cost of having virtual destructors.

More functionality will likely be added later, as the need for it arises.